fix: dashboard audit — 11 critical/moderate fixes
Critical:
- C1: Fix pantry macro display (protein_g/carbs_g/fats_g field names)
- C2: Remove Notion dependency from add_meal_from_search (local DB only)
- C3: Fix JS payload field names in addFood()
- C5: Change Meal.kcal int->float to match agent output

Moderate:
- M1: Replace all hardcoded values in garden.html with dynamic data
- M3: Macro % now uses caloric contribution (P*4, C*4, F*9)
- M4: Add Sign Out button to profile page
- M5: Sync target_kcal from user profile instead of global default

Security:
- U5: Fix XSS in home.html meal names via |tojson filter
- U8: Remove unused Notion import from get_common_context

+++ app/database.py
@@ -70,7 +70,7 @@ class Meal(SQLModel, table=True):
     telegram_user_id: int = Field(index=True)
     date: str = Field(index=True)  # YYYY-MM-DD format
     name: str
-    kcal: int = Field(default=0)
+    kcal: float = Field(default=0.0)
     protein_g: float = Field(default=0.0)
     carbs_g: float = Field(default=0.0)
     fats_g: float = Field(default=0.0)
